Fix React Native tests
Summary: At Facebook we do not install the cli or the packages around it. I added some workaround for that in D15044762 but failed to run Jest without cache to test it and it seems like it didn't run on sandcastle properly either (wtf?). Either way, this adds back the `ios` and `android` reducers for `hasteImpl` to run the tests properly again.

Reviewed By: rubennorte

Differential Revision: D15063152

fbshipit-source-id: 2f1fefe664f02d5f2f5e65b75ac1e0c5b813a343