Disable localhostBundleURL check in testBundleURL, without breaking stable
Summary:
This test was disabled in e106112202f9b805162d2b814ddf14ab29a0f97b to allow unit tests to run on Facebook's internal CI. The change was reverted in 1f6de88230ffd076e57838dbad7b48b71b1be6a0 because another internal test was broken when it found that mainBundleURL() was not getting called.

In this commit, I've commented out the actual piece of code that would cause unit tests to fail in Facebook's internal CI, without removing the call to mainBundleURL(). The localhostBundleURL() method is called elsewhere in the file, so commenting it out here should not cause any issues.

Changelog:
[iOS] [Changed] - Disable testBundleURL test.

Reviewed By: cpojer

Differential Revision: D15049238

fbshipit-source-id: da3a393922f2190b423980cac5ab54df5e7e3e41