Check for type before force conversion
Summary:
If you try to use ART components with fabric, their transform is causes crash (stack trace P108110438). You can see this crash if you use RCTVideo component in fabric and tap video to reveal slider.

The cause of crash is mismatch of formats between what is expected for transform and what is being sent.

In this diff we add a check to see whether the configuration is of expected type.

Reviewed By: shergin

Differential Revision: D17181838

fbshipit-source-id: c4f3c920281a2e7f58ff0ffe1d0ec2af8249a16c